// employee-app/src/types/employee.types.ts
export type MyProfile = {
  emp_id: number;
  emp_code: string;
  first_name: string;
  last_name: string;
  email_1: string | null;
  email_2: string | null;
  mobile_no_1: string | null;
  mobile_no_2: string | null;
  address: string | null;
  gender: string | null;
  nic: string | null;
  profile_photo: string | null;
  joining_date: string | null;
  employment_status: string | null;
  designation: string | null;
  role_id: number;
  role_name: string | null;
};